requirements.txt
pandas
matplotlib
selenium
python-dotenv
clarifai-grpc